requirements.txt
boto3==1.34.119
pymongo==4.7.2
python-dotenv==1.0.1
joblib==1.4.2
botocore~=1.34.123
aioboto3